+++ src/main/java/inu/spp/cryptocoffee/domain/company/entity/CompanyEntity.java
@@ -1,4 +1,4 @@
-package inu.spp.cryptocoffee.domain.company.enttiy;
+package inu.spp.cryptocoffee.domain.company.entity;
 
 import inu.spp.cryptocoffee.domain.company.dto.Address;
 import inu.spp.cryptocoffee.global.base.BaseTimeEntity;
